app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
pydantic==2.9.2
psycopg2-binary==2.9.9
